bug(SIDM-4848): Allow to turn SSO off without conf (#483)
* bug(sidm-4848): refactor sso config to allow empty SSO config when disabled

* bug(sidm-4848): SSOAuthenticationHandler conditional on sso enabled

* bug(sidm-4848): create RestTemplateConfiguration
